refactor: reorganizando as contants.py das pipeline para usarem settings
