+++ pipelines/acolherio/flows.py
@@ -1,12 +1,10 @@
 from prefect import flow
 from pipelines.tasks import run_dbt_models
-from pipelines.utils.settings import BasePipelineSettings
+from pipelines.acolherio.constants import settings
 
 
 @flow(name="dbt_acolherio_rma")
 def acolherio_flow() -> None:
-    # Apenas instanciar a classe base já configura a autenticação GCP e o MODE
-    settings = BasePipelineSettings()
     
     run_dbt_models("stg_filtro_evolucao")
